Refactor: Decouple `AimsCoachingHandler` by introducing dedicated service dependencies and updating related tests
