+++ app/services/classifier_service.py
@@ -192,7 +192,7 @@ class ClassifierService:
     ) -> Dict[str, Any]:
         """Call Gemini to detect if the session has reached a natural conclusion.
 
-        TODO: Wire this into AimsCoachingHandler._check_end_game() as an LLM-based
+        TODO: Wire this into AimsEndgameService.check() as an LLM-based
         complement/replacement to the heuristic EndGameDetector. Currently prep work only.
         """
         prompt = AimsPromptBuilder.build_endgame_detector_prompt(
