+++ tests/test_aims_improvements.py
@@ -72,7 +72,7 @@ def test_multisentence_rapport_q_is_smalltalk(aims_mapping):
 # Phase flexibility tests (Fix 5)
 # ---------------------------------------------------------------------------
 
-from app.services.aims_coaching_handler import AimsCoachingHandler
+from app.services.aims_state_service import AimsStateService
 
 
 def _make_state(phase="Secure", concerns=None):
@@ -85,69 +85,60 @@ def _make_state(phase="Secure", concerns=None):
     }
 
 
-def _handler_instance():
-    """Minimal handler just to call _update_observational_state."""
+def _state_service():
+    """Minimal state service for AIMS state transition tests."""
     import logging
-    return AimsCoachingHandler(
-        memory_store={},
-        vertex_config={
-            "project_id": "p", "region": "r", "vertex_location": "r",
-            "model_id": "m", "model_fallbacks": [],
-            "temperature": 0.0, "max_tokens": 256, "client_cls": None,
-        },
-        memory_config={"enabled": False, "max_turns": 10},
-        logger=logging.getLogger("test"),
-    )
+    return AimsStateService(logger=logging.getLogger("test"))
 
 
 def test_mirror_returns_phase_to_inquire_mirror_from_secure():
     """A Mirror step detected while in Secure phase should cycle back to InquireMirror."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="Secure")
-    h._update_observational_state(state, "Mirror", ["Mirror"])
+    h.update_observational_state(state, "Mirror", ["Mirror"])
     assert state["phase"] == "InquireMirror"
 
 
 def test_inquire_returns_phase_to_inquire_mirror_from_secure():
     """An Inquire step detected while in Secure phase should cycle back to InquireMirror."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="Secure")
-    h._update_observational_state(state, "Inquire", ["Inquire"])
+    h.update_observational_state(state, "Inquire", ["Inquire"])
     assert state["phase"] == "InquireMirror"
 
 
 def test_secure_stays_in_secure_when_all_concerns_mirrored():
     """Secure step with all concerns mirrored should keep phase as Secure."""
-    h = _handler_instance()
+    h = _state_service()
     concerns = [{"desc": "x", "topic": "t", "is_mirrored": True, "is_secured": False}]
     state = _make_state(phase="InquireMirror", concerns=concerns)
-    h._update_observational_state(state, "Secure", ["Secure"])
+    h.update_observational_state(state, "Secure", ["Secure"])
     assert state["phase"] == "Secure"
 
 
 def test_secure_stays_in_inquire_mirror_when_unmirrored_concerns_remain():
     """Secure step with unmirrored concerns should NOT advance phase to Secure."""
-    h = _handler_instance()
+    h = _state_service()
     concerns = [
         {"desc": "x", "topic": "t1", "is_mirrored": True, "is_secured": True},
         {"desc": "y", "topic": "t2", "is_mirrored": False, "is_secured": False},
     ]
     state = _make_state(phase="InquireMirror", concerns=concerns)
-    h._update_observational_state(state, "Secure", ["Secure"])
+    h.update_observational_state(state, "Secure", ["Secure"])
     assert state["phase"] == "InquireMirror"
 
 
 def test_mirror_plus_inquire_returns_phase_from_secure():
     """Mirror+Inquire compound step should return to InquireMirror from Secure."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="Secure")
-    h._update_observational_state(state, "Mirror+Inquire", ["Mirror", "Inquire"])
+    h.update_observational_state(state, "Mirror+Inquire", ["Mirror", "Inquire"])
     assert state["phase"] == "InquireMirror"
 
 
 def test_mirror_secure_inquire_scalar_marks_all_components():
     """A scalar-only Mirror+Secure+Inquire payload should still update concern state."""
-    h = _handler_instance()
+    h = _state_service()
     mem = {
         "character": None,
         "aims_state": _make_state(
@@ -164,7 +155,7 @@ def test_mirror_secure_inquire_scalar_marks_all_components():
     }
     cls = {"step": "Mirror+Secure+Inquire", "score": 3, "reasons": [], "tips": []}
 
-    h._update_aims_state(
+    h.update(
         mem,
         cls,
         "You're worried about side effects. Serious side effects are rare. What else is on your mind?",
@@ -182,7 +173,7 @@ def test_mirror_secure_inquire_scalar_marks_all_components():
 
 def test_mirror_secure_inquire_can_resolve_multiple_explicit_concerns():
     """A blended turn mentioning two concerns should mirror and secure both."""
-    h = _handler_instance()
+    h = _state_service()
     mem = {
         "character": None,
         "aims_state": _make_state(
@@ -205,7 +196,7 @@ def test_mirror_secure_inquire_can_resolve_multiple_explicit_concerns():
     }
     cls = {"step": "Mirror+Secure+Inquire", "score": 3, "reasons": [], "tips": []}
 
-    h._update_aims_state(
+    h.update(
         mem,
         cls,
         (
@@ -225,7 +216,7 @@ def test_mirror_secure_inquire_can_resolve_multiple_explicit_concerns():
 
 def test_secure_inquire_scalar_marks_secure_without_premature_secure_warning():
     """Secure+Inquire should secure an already mirrored concern without Secure-only warning."""
-    h = _handler_instance()
+    h = _state_service()
     mem = {
         "character": None,
         "aims_state": _make_state(
@@ -242,7 +233,7 @@ def test_secure_inquire_scalar_marks_secure_without_premature_secure_warning():
     }
     cls = {"step": "Secure+Inquire", "score": 3, "reasons": [], "tips": []}
 
-    h._update_aims_state(
+    h.update(
         mem,
         cls,
         "Side effects are usually mild and brief. What else would help you decide?",
@@ -264,51 +255,51 @@ def test_secure_inquire_scalar_marks_secure_without_premature_secure_warning():
 
 def test_secure_before_mirror_first_time_gives_standard_feedback():
     """First Secure-before-mirror should give standard feedback."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="InquireMirror", concerns=[
         {"desc": "x", "topic": "trust", "is_mirrored": False, "is_secured": False},
     ])
     cls = {"step": "Secure", "score": 2, "reasons": [], "tips": []}
-    h._apply_coaching_guidance(cls, "Secure", state, "The data shows...", "I don't trust pharma")
+    h.apply_coaching_guidance(cls, "Secure", state, "The data shows...", "I don't trust pharma")
     assert "reflecting" in cls["reasons"][0].lower() or "mirroring" in cls["reasons"][0].lower()
     assert state.get("recent_coaching") == ["secure_before_mirror"]
 
 
 def test_secure_before_mirror_second_time_escalates_with_topic():
     """Second repetition should name the unmirrored concern topic."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="InquireMirror", concerns=[
         {"desc": "x", "topic": "trust", "is_mirrored": False, "is_secured": False},
     ])
     state["recent_coaching"] = ["secure_before_mirror"]  # simulate 1 prior
     cls = {"step": "Secure", "score": 2, "reasons": [], "tips": []}
-    h._apply_coaching_guidance(cls, "Secure", state, "Studies show...", "I don't trust pharma")
+    h.apply_coaching_guidance(cls, "Secure", state, "Studies show...", "I don't trust pharma")
     assert "trust" in cls["reasons"][0].lower()
     assert "still" in cls["reasons"][0].lower() or "hasn't" in cls["reasons"][0].lower()
 
 
 def test_secure_before_mirror_third_time_escalates_to_pattern():
     """Third+ repetition should produce a pattern-level escalation."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="InquireMirror", concerns=[
         {"desc": "x", "topic": "trust", "is_mirrored": False, "is_secured": False},
     ])
     state["recent_coaching"] = ["secure_before_mirror", "secure_before_mirror"]
     cls = {"step": "Secure", "score": 2, "reasons": [], "tips": []}
-    h._apply_coaching_guidance(cls, "Secure", state, "Evidence says...", "I don't trust pharma")
+    h.apply_coaching_guidance(cls, "Secure", state, "Evidence says...", "I don't trust pharma")
     assert "3" in cls["reasons"][0]  # should mention the count
     assert "trust" in cls["reasons"][0].lower()
 
 
 def test_secure_after_mirroring_resets_coaching_counter():
     """Securing after all concerns mirrored should reset the coaching counter."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="Secure", concerns=[
         {"desc": "x", "topic": "trust", "is_mirrored": True, "is_secured": False},
     ])
     state["recent_coaching"] = ["secure_before_mirror", "secure_before_mirror"]
     cls = {"step": "Secure", "score": 2, "reasons": [], "tips": []}
-    h._apply_coaching_guidance(cls, "Secure", state, "The data is clear...", "")
+    h.apply_coaching_guidance(cls, "Secure", state, "The data is clear...", "")
     assert state["recent_coaching"] == []  # counter should be reset
 
 
@@ -318,31 +309,31 @@ def test_secure_after_mirroring_resets_coaching_counter():
 
 def test_detect_trust_style_analytical():
     """Character with analytical keywords should return 'analytical'."""
-    assert AimsCoachingHandler._detect_trust_style(
+    assert AimsStateService.detect_trust_style(
         "Ethan is analytical, values peer-reviewed evidence."
     ) == "analytical"
 
 
 def test_detect_trust_style_default():
     """Character without analytical keywords should return 'default'."""
-    assert AimsCoachingHandler._detect_trust_style(
+    assert AimsStateService.detect_trust_style(
         "Sarah is a caring mother who values her child's safety."
     ) == "default"
 
 
 def test_detect_trust_style_none():
     """None character should return 'default'."""
-    assert AimsCoachingHandler._detect_trust_style(None) == "default"
+    assert AimsStateService.detect_trust_style(None) == "default"
 
 
 def test_analytical_persona_gets_reasoning_tip():
     """An analytical persona should get 'validate reasoning' tip instead of emotional one."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="InquireMirror", concerns=[
         {"desc": "risk level", "topic": "side_effects", "is_mirrored": False, "is_secured": False},
     ])
     cls = {"step": "Secure", "score": 2, "reasons": [], "tips": []}
-    h._apply_coaching_guidance(
+    h.apply_coaching_guidance(
         cls, "Secure", state, "The data shows...", "That risk seems high",
         character="Ethan is analytical, values peer-reviewed evidence.",
     )
@@ -353,12 +344,12 @@ def test_analytical_persona_gets_reasoning_tip():
 
 def test_default_persona_gets_emotional_tip():
     """A default persona should get the standard emotional tip."""
-    h = _handler_instance()
+    h = _state_service()
     state = _make_state(phase="InquireMirror", concerns=[
         {"desc": "scared", "topic": "side_effects", "is_mirrored": False, "is_secured": False},
     ])
     cls = {"step": "Secure", "score": 2, "reasons": [], "tips": []}
-    h._apply_coaching_guidance(
+    h.apply_coaching_guidance(
         cls, "Secure", state, "The data shows...", "I'm scared",
         character="Sarah is a caring mother.",
     )
